+++ multiqc/modules/somalier/somalier.py
@@ -573,7 +573,7 @@ class MultiqcModule(BaseMultiqcModule):
         if len(data) > 0:
             pconfig = {
                 "id": "somalier_sex_check_plot",
-                "title": "Somalier: Sample Predicted Sex",
+                "title": "Somalier: Provided Sex vs. Scaled X Depth",
                 "xlab": "Sex from pedigree",
                 "ylab": "Scaled mean depth on X",
                 "categories": ["Female", "Male", "Unknown"],
@@ -581,8 +581,8 @@ class MultiqcModule(BaseMultiqcModule):
 
             self.add_section(
                 name="Sex",
-                description="Predicted sex against scaled depth on X",
-                helptext="Higher values of depth, low values suggest male.",
+                description="Provided sex against scaled depth on X",
+                helptext="Higher depth values suggest female, lower values suggest male.",
                 anchor="somalier-sexcheck",
                 plot=scatter.plot(data, pconfig),
             )